+++ packages/help-center/src/components/notices.tsx
@@ -3,13 +3,9 @@ import { localizeUrl } from '@automattic/i18n-utils';
 import { useCanConnectToZendeskMessaging } from '@automattic/zendesk-client';
 import { Button } from '@wordpress/components';
 import { useDispatch } from '@wordpress/data';
-import { createInterpolateElement } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
-import { Icon, info } from '@wordpress/icons';
 import { useEffect } from 'react';
-import { useLocation, useNavigate } from 'react-router-dom';
 import { useFeatureConfig, useHelpCenterContext } from '../contexts/HelpCenterContext';
-import { useSupportStatus } from '../data/use-support-status';
 import useChatStatus from '../hooks/use-chat-status';
 import './notices.scss';
 import { HELP_CENTER_STORE } from '../stores';
@@ -58,43 +54,6 @@ export const BlockedZendeskNotice: React.FC = () => {
 				>
 					{ __( 'Learn more.', __i18n_text_domain__ ) }
 				</Button>
-			</p>
-		</div>
-	);
-};
-
-export const EmailFallbackNotice: React.FC = () => {
-	const navigate = useNavigate();
-	const { search } = useLocation();
-	const { data: supportStatus } = useSupportStatus();
-	const params = new URLSearchParams( search );
-	params.set( 'wapuuFlow', 'true' );
-	const url = '/contact-form?' + params.toString();
-
-	const title = supportStatus?.eligibility?.is_chat_restricted
-		? __( 'Live chat is currently unavailable.', __i18n_text_domain__ )
-		: __( 'Live chat is temporarily unavailable for scheduled maintenance.', __i18n_text_domain__ );
-
-	const message = __(
-		'Please reach out via <email>email</email> if you need assistance.',
-		__i18n_text_domain__
-	);
-
-	return (
-		<div className="help-center__notice email-fallback-notice">
-			<Icon icon={ info } className="help-center__notice-icon" />
-			<p>
-				<strong>{ title }</strong>
-				&nbsp;
-				{ createInterpolateElement( message, {
-					email: (
-						<Button
-							variant="link"
-							className="help-center__notice-link"
-							onClick={ () => navigate( url ) }
-						/>
-					),
-				} ) }
 			</p>
 		</div>
 	);
